requirements.txt
gunicorn
flask
torch
torchvision
Pillow
